requirements.txt
flask>=2.0.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
xgboost>=1.6.0
lightgbm>=3.3.0
tensorflow>=2.10.0
librosa>=0.9.0
praat-parselmouth>=0.4.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
scipy>=1.8.0
joblib>=1.1.0
werkzeug>=2.0.0
flask-login>=0.6.0
flask-bcrypt>=1.0.0
psycopg2-binary>=2.9.0
flask-wtf>=1.1.0
wtforms>=3.0.0
flask-mail>=0.9.1 